enh: Provide `getCurrentSkinTone` and `setCurrentSkinTone` functions and use in `EmojiSearch`
