+++ src/functions/index.js
@@ -1,3 +1,3 @@
 export * from './a11y/index.ts'
-export * from './emoji/index.js'
+export * from './emoji/index.ts'
 export { default as usernameToColor } from './usernameToColor/index.js'
